新增預覽圖功能及accessoryButton

+++ ShareExtension/ViewController/ShareViewController.swift
@@ -14,6 +14,7 @@ class ShareViewController: SLComposeServiceViewController {
     @IBOutlet weak var myImageView: UIImageView!
     let suiteName = "group.maxkit.fred.ShareExtensionTest"
     let manager = FavoriteManager.shareInstance
+    
     override func isContentValid() -> Bool {
         // Do validation of contentText and/or NSExtensionContext attachments here
         return true
@@ -21,7 +22,6 @@ class ShareViewController: SLComposeServiceViewController {
     
 
     override func didSelectPost() {
-        
         if let inputItems = extensionContext!.inputItems as? [NSExtensionItem] {
             
             for inputItem in inputItems {
@@ -33,15 +33,16 @@ class ShareViewController: SLComposeServiceViewController {
                         print("attachment: \(attachment)")
                         
                         switch (true) {
+                            
                         case attachment.hasItemConformingToTypeIdentifier(ContentType.publicJpeg.rawValue):
                             self.retriveAttachment(attachment: attachment, type: .publicJpeg)
-                        
-                        case attachment.hasItemConformingToTypeIdentifier(ContentType.publicUrl.rawValue):
                             
+                        case attachment.hasItemConformingToTypeIdentifier(ContentType.publicUrl.rawValue):
                             switch (true) {
                             case attachment.hasItemConformingToTypeIdentifier(ContentType.publicZip_archive.rawValue):
                                 
                                 switch (true) {
+                                    
                                 case attachment.hasItemConformingToTypeIdentifier(ContentType.comAppleIworkPagesSffpages.rawValue):
                                     self.retriveAttachment(attachment: attachment, type: .comAppleIworkPagesSffpages)
 
@@ -55,9 +56,6 @@ class ShareViewController: SLComposeServiceViewController {
                             case attachment.hasItemConformingToTypeIdentifier(ContentType.publicPng.rawValue):
                                 self.retriveAttachment(attachment: attachment, type: .publicPng)
                                 
-                            case attachment.hasItemConformingToTypeIdentifier(ContentType.publicPlainText.rawValue):
-                                self.retriveAttachment(attachment: attachment, type: .publicPlainText)
-                                
                             case attachment.hasItemConformingToTypeIdentifier(ContentType.publicMpeg4.rawValue):
                                 self.retriveAttachment(attachment: attachment, type: .publicMpeg4)
                                 
@@ -65,17 +63,16 @@ class ShareViewController: SLComposeServiceViewController {
                                 self.retriveAttachment(attachment: attachment, type: .publicUrl)
                             }
                             
+                        case attachment.hasItemConformingToTypeIdentifier(ContentType.publicPlainText.rawValue):
+                            self.retriveAttachment(attachment: attachment, type: .publicPlainText)
+                            
                         default:
                             print("default attachment: \(attachment)")
                         }
-
                     }
                 }
             }
         }
-        
-        
-
         print("Done")
         self.extensionContext!.completeRequest(returningItems: [], completionHandler: nil)
     }
@@ -89,43 +86,37 @@ class ShareViewController: SLComposeServiceViewController {
     // MARK: - To get info of attachment
     func retriveAttachment(attachment: NSItemProvider, type: ContentType) {
         attachment.loadItem(forTypeIdentifier: type.rawValue, options: nil) {
+            
             (data, error) in
-            if error != nil { print(error!.localizedDescription) }
+            print("data: \(data)")
+            
+            if error != nil { print(error) }
+            
             if let url = data as? URL {
                 
                 // Enable permission
                 url.startAccessingSecurityScopedResource()
-                
-                let fileObject = FileObject(name: url.lastPathComponent, type: type, url: url, size: self.getSize(atPath: url.path)!)
-                
-                print("fileObject: \(fileObject)")
-                
-                self.manager.setFavoriteData(fileObject)
-                
+
+                self.setToObjects(name: url.lastPathComponent, type: type, url: url, size: self.getSize(atPath: url.path)!)
+        
                 // METHOD: FileManager
                 if let shareUrl = FileManager.default.containerURL(forSecurityApplicationGroupIdentifier: self.suiteName) {
                     do {
                         let tmpData = try Data(contentsOf: url)
+                        let tmpDataPath = shareUrl.appendingPathComponent(url.lastPathComponent)
                         
-                        let imagePath = shareUrl.appendingPathComponent(url.lastPathComponent)
-                        print("imagePath: \(imagePath)")
+                        /*
+                        // Remove duplicated item
+                        if FileManager().fileExists(atPath: imagePath.path) { try! FileManager().removeItem(at: imagePath) }
+                        */
                         
-                        if FileManager().fileExists(atPath: imagePath.path) {
-                            print("imagePath existed!")
-                            try! FileManager().removeItem(at: imagePath)
-                            print("remove successed")
-                        }
-                        
-                        try! tmpData.write(to: imagePath)
+                        try! tmpData.write(to: tmpDataPath)
                         print("write successed")
-                           
                         
                     } catch {
                         print(error)
                     }
-                    
-                    
-                    
+   
                 }
                 
                 
@@ -148,26 +139,38 @@ class ShareViewController: SLComposeServiceViewController {
                 // Disable permission
                 url.stopAccessingSecurityScopedResource()
                 
+            } else if let str = data as? String {
                 
+                // No url and size exists, if is simple text
+                self.setToObjects(name: str, type: type, url: nil, size: 0)
+                
+            } else {
+                print("Can not unwrapping to type.")
+                print("data: \(data)")
             }
         }
         
     }
     
+    func setToObjects(name: String, type: ContentType, url:  URL?, size: UInt64) {
+        let fileObject = FileObject(name: name, type: type, url: url, size: size)
+        self.manager.setFavoriteData(fileObject)
+        print("fileObject: \(fileObject)")
+        
+    }
     
     func getSize(atPath path: String) -> UInt64? {
-        
         do {
             let attr = try FileManager.default.attributesOfItem(atPath: path)
             let fileSize = attr[FileAttributeKey.size] as! UInt64
-            
             return fileSize
+            
         } catch {
             
-            // Return 0, if a web url
+            // Return 0, if is web url
             return 0
+            
         }
     }
-    
     
 }
